feat: move task editor card drag and drop context into separate wrapper component
